Trick or Treaters
Halloween map now features trick-or-treaters who try to reach houses before you. If they stay on a house for a full turn they close the door and you can no longer get the candy from that house. Astral shard is now granted if you get 4 houses (the majority) instead of if you get all of them.